+++ packages/lib/package/components/MazPicker/MazPickerHeader.vue
@@ -51,16 +51,13 @@
 </template>
 
 <script lang="ts" setup>
-  import { computed, PropType, ref, watch } from 'vue'
-  import { Color } from '../types'
+  import { computed, type PropType, ref, watch } from 'vue'
+
+  import type { Color } from '../types'
   import { date, capitalize } from '@package/filters'
-  import { PickerValue } from './types'
-  import {
-    cloneDate,
-    DateTimeFormatOptions,
-    getCurrentDateForTimeValue,
-    isBigger,
-  } from './utils'
+  import type { PickerValue } from './types'
+  import type { DateTimeFormatOptions } from './utils'
+  import dayjs from 'dayjs'
 
   const props = defineProps({
     modelValue: {
@@ -72,48 +69,26 @@
     noShortcuts: { type: Boolean, required: true },
     double: { type: Boolean, required: true },
     hasDate: { type: Boolean, required: true },
-    time: { type: Boolean, required: true },
+    hasTime: { type: Boolean, required: true },
     formatterOptions: {
       type: Object as PropType<DateTimeFormatOptions>,
       required: true,
     },
-    currentDate: { type: Date, required: true },
+    calendarDate: { type: String, required: true },
   })
 
-  const getCurrentDateTmp = () => {
-    return typeof props.modelValue === 'object'
-      ? cloneDate(
-          props.modelValue.start
-            ? new Date(props.modelValue.start)
-            : new Date(),
-        )
-      : cloneDate(
-          props.modelValue
-            ? props.hasDate
-              ? new Date(props.modelValue)
-              : getCurrentDateForTimeValue(props.modelValue)
-            : new Date(),
-        )
-  }
-
-  const currentDateTmp = ref<Date>(getCurrentDateTmp())
-
-  const selectedDate = computed<string | undefined>(() => {
-    if (typeof props.modelValue === 'object') {
-      return props.modelValue.start
-    }
-    return props.modelValue
-  })
+  const refDate = computed(() =>
+    typeof props.modelValue === 'string'
+      ? props.modelValue
+      : props.modelValue?.start,
+  )
 
   const transitionName = ref<'maz-slidevnext' | 'maz-slidevprev'>(
     'maz-slidevnext',
   )
 
   const year = computed(() => {
-    if (
-      typeof props.modelValue === 'object' &&
-      (props.modelValue.start || props.modelValue.end)
-    ) {
+    if (typeof props.modelValue === 'object') {
       return `${
         props.modelValue.start
           ? date(props.modelValue.start, props.locale, {
@@ -189,36 +164,39 @@
   )
 
   const timeValue = computed(() => {
-    return date(currentDateTmp.value, props.locale, {
-      timeStyle: 'short',
-      timeZone: props.formatterOptions.timeZone,
-      hour12: props.formatterOptions.hour12,
-    })
+    return refDate.value
+      ? date(refDate.value, props.locale, {
+          timeStyle: 'short',
+          timeZone: props.formatterOptions.timeZone,
+          hour12: props.formatterOptions.hour12,
+        })
+      : undefined
   })
 
-  const timeArray = computed(() => (props.time ? [timeValue.value] : undefined))
+  const timeArray = computed(() =>
+    props.hasTime ? [timeValue.value] : undefined,
+  )
 
   watch(
     () => props.modelValue,
-    () => {
-      if (selectedDate.value) {
-        transitionName.value = isBigger(
-          currentDateTmp.value,
-          new Date(selectedDate.value),
-        )
-          ? 'maz-slidevprev'
-          : 'maz-slidevnext'
-        currentDateTmp.value = props.hasDate
-          ? cloneDate(new Date(selectedDate.value))
-          : getCurrentDateForTimeValue(selectedDate.value)
-      }
+    (modelValue, oldModelValue) => {
+      const currentValue =
+        typeof modelValue === 'object' ? modelValue.start : modelValue
+      const oldValue =
+        typeof oldModelValue === 'object' ? oldModelValue.start : oldModelValue
+
+      transitionName.value =
+        dayjs(currentValue).isAfter(oldValue, 'date') ||
+        dayjs(currentValue).isSame(oldValue, 'date')
+          ? 'maz-slidevnext'
+          : 'maz-slidevprev'
     },
   )
 </script>
 
 <style lang="postcss" scoped>
   .m-picker-header {
-    @apply maz-z-1 maz-flex maz-justify-between maz-space-y-1 maz-p-2;
+    @apply maz-z-1 maz-flex maz-justify-between maz-space-y-1 maz-p-2 maz-font-normal;
 
     &__year-transition {
       @apply maz-flex maz-h-5 maz-items-center maz-overflow-hidden maz-leading-3;
@@ -301,10 +279,6 @@
 
     &.--black {
       @apply maz-bg-black maz-text-black-contrast;
-
-      & .m-picker-header__year-text {
-        @apply maz-text-muted;
-      }
     }
 
     &.--white {