+++ packages/lib/package/components/MazSpinner.vue
@@ -23,7 +23,7 @@
 </script>
 
 <script lang="ts" setup>
-  import { PropType } from 'vue'
+  import type { PropType } from 'vue'
   import type { Color } from './types'
 
   defineProps({