+++ packages/lib/package/directives/v-zoom-img/zoom-img.handler.ts
@@ -1,4 +1,4 @@
-import { DirectiveBinding } from 'vue'
+import type { DirectiveBinding } from 'vue'
 import { style } from './style'
 import { svgs } from './svgs'