+++ packages/lib/vitest.config.ts
@@ -1,27 +1,34 @@
+/// <reference types="vitest" />
+
 import { defineConfig } from 'vitest/config'
 import Vue from '@vitejs/plugin-vue'
+
 import { resolve } from 'path'
 
 const projectRoot = resolve(__dirname)
 
 export default defineConfig({
   plugins: [Vue()],
+  server: {
+    port: 1000,
+  },
   test: {
     globals: true,
     environment: 'jsdom',
     coverage: {
-      enabled: true,
+      all: true,
       excludeNodeModules: true,
       reporter: ['clover', 'html'],
-      include: ['package/**/*'],
+      include: ['package'],
       exclude: ['package/components_tmp/**'],
-      all: true,
+      extension: ['.js', '.ts', '.vue'],
     },
   },
   resolve: {
     alias: {
       '@package': resolve(projectRoot, 'package'),
-      '@unit': resolve(projectRoot, 'tests/unit'),
+      '@components': resolve(projectRoot, 'package/components'),
+      '@tests': resolve(projectRoot, 'tests'),
     },
   },
 })