+++ packages/testing/stylelint.config.js
@@ -3,6 +3,8 @@ module.exports = {
     'stylelint-config-standard',
     'stylelint-config-prettier',
     'stylelint-config-recommended-vue',
+    'stylelint-config-standard-scss',
+    'stylelint-config-tailwindcss',
   ],
   // add your custom config here
   // https://stylelint.io/user-guide/configuration